fix(declarationHistory): split DeclarationHistoryPage into server + client components
Move pagination state and tRPC query from DeclarationHistoryPage (now a
server component) into new HistoryListSection client component to avoid
pulling server-only modules into the client bundle.
